Move GridControl into general devices/
