+++ src/main/java/net/perkowitz/issho/hachi/Hachi.java
@@ -251,4 +251,5 @@ public class Hachi {
         return modules;
     }
 
+
 }
