+++ src/main/java/net/perkowitz/issho/hachi/Knobby.java
@@ -1,0 +1,10 @@
+package net.perkowitz.issho.hachi;
+
+/**
+ * Created by optic on 11/16/16.
+ */
+public class Knobby {
+
+
+
+}
